Fixed bad conversions

+++ src/protos/lightheader.h
@@ -143,7 +143,7 @@ struct FlashData {
         ledcool_current (ToOptional(message->GetUInt32Array(2))),
         ledwarm_current (ToOptional(message->GetUInt32Array(3))),
         flash_duration (ToOptional(message->GetUInt32Array(4))),
-        offset_duration (ToOptional(message->GetInt64Array(5)))
+        offset_duration (ToOptional(message->GetInt32Array(5)))
     {}
 };
 
@@ -164,7 +164,7 @@ struct AFDebugInfo {
 
     AFDebugInfo(picoproto::Message* message) :
         focus_achieved (ToOptional(message->GetBoolArray(1))),
-        image_focal_length (ToOptional(message->GetInt64Array(2)))
+        image_focal_length (ToOptional(message->GetInt32Array(2)))
     {}
 };
